requirements.txt
''' 
set up virtual environment

# Linux
sudo apt-get install python3-venv    # If needed
python3 -m venv .venv
source .venv/bin/activate

# macOS
python3 -m venv .venv
    source .venv/bin/activate

# Windows
py -3 -m venv .venv
.venv\scripts\activate
'''

# Install dependencies by `pip install -r requirements.txt`

django
django-bootstrap-v5
django-simplemde
Faker
sqlparse
setuptools
markdown
django-taggit
